Extracts getIfPathToSchemas
