+++ samples/client/3_1_0_unit_test/java/src/main/java/org/openapijsonschematools/client/schemas/validation/KeywordValidator.java
@@ -7,11 +7,4 @@ public interface KeywordValidator {
     @Nullable PathToSchemasMap validate(
         ValidationData data
     ) throws ValidationException;
-
-    default PathToSchemasMap getIfPathToSchemas(
-        @Nullable Object arg,
-        ValidationMetadata validationMetadata
-    ) {
-        return new PathToSchemasMap();
-    }
 }
