+++ src/main/resources/java/src/main/java/packagename/schemas/validation/KeywordValidator.hbs
@@ -7,11 +7,4 @@ public interface KeywordValidator {
     @Nullable PathToSchemasMap validate(
         ValidationData data
     ) throws ValidationException;
-
-    default PathToSchemasMap getIfPathToSchemas(
-        @Nullable Object arg,
-        ValidationMetadata validationMetadata
-    ) {
-        return new PathToSchemasMap();
-    }
 }
